Update local-browser-w.service
Add a delay to startup of the browser to see if it solves the scaling issue upon startup. What happens currently is sometimes the scaling is not set when the local browser is started from this service file which starts weston if it is not running. It has not happened when weston is already running. This is a hack to see if it is a timing issue and solves the issue where scaling does not seem to apply every time.