fixed NAT
Previous 'fix' when switching to connman was to move teh dnsmasq port yo 5300. I assum,e this worked at one point, but recently noted that NAT was broken. There was a port interaction problem between connman and dnsmasq on port 53, so we moved dnsmasq to 5300. This then messed up DNS routing. Removing these lines restores NAT operation and does not appear to interfere with connman.