Update weston.service
Weston can now be started as user http without resorting to the work-around of chowning the /run/weston structure after starting. I believe this is a bug fix in systemd as this workaround was previously required to run which caused weston to be run as root. weston now runs happily as user http