Update weston.service
This is a work around because weston cannot run without systemd-logind running and a user with a password. We have systemd-logind masked so that no-one other than root can log in. Weston is designed to be run as a user. Runeaudio has no users and, therefore cannot run weston as a user. The current workaround to this is to run weston as root, then change the ownership of /run/weston to http:http so any applications running under weston can run as http:http. If weston changes or a different compositor can be run, this can be revisited. As it is, weston is the way we can run a local browser w/o X11.