Delete app/config/defaults/etc/tmpfiles.d/connman_resolvconf.conf
This file made sure that /rubn/connman was created and also that resolv.conf pointed to /rub/connman/resolv.conf. We have changed to systemd-resolved and this is no longer necessary. A new file called connman.conf ensures that /run/connman is created and not this link to the /etc/resolv.conf