Removed non-UI code from the PHP opcache
It seems that adding CLI PHP code to the PHP opcache achieves nothing except in very special circumstances. Now removed, should give a few more bytes of space in main memory. PHP opcache is set up with a high limit, but it only uses memory as required below this level.