revised disconnect-delete case in wrk_netconfig still does not work at all. Nothing happens. no log errors.
this case should remove the networkd config and restart networkd
this removes any connection static or dhcp. Oncew this is done, it
then should remove the wpa_cli/iwd config. This is best done via
the cli app versus editing the config as it will then update on
the fly and not require a restart of iwd/wpa_supplicant@wlan0
to register the removal of an ssid...